+++ lib/pgchief/config/s3.rb
@@ -13,7 +13,7 @@ module Pgchief
         :s3_key,
         :s3_secret,
         :s3_region,
-        :s3_path_prefix
+        :s3_objects_path
 
       PREFIX_REGEX = %r{\As3://(?<bucket>(\w|-)*)/(?<path>(\w|/)*/)\z}
 
@@ -44,14 +44,14 @@ module Pgchief
           s3_key,
           s3_secret,
           s3_region,
-          s3_path_prefix
+          s3_objects_path
         ].none?(&:nil?)
       end
 
       private
 
       def s3_match
-        @s3_match ||= s3_path_prefix.match(PREFIX_REGEX)
+        @s3_match ||= s3_objects_path.match(PREFIX_REGEX)
       end
     end
   end